Update README.md
# Python-Challenge

THis is Python code, prepared by Edris Gemtessa at University of minnesota.
this code analysis ByBank financial analysis and ByPoll election analysis. as the following:
  The total number of months included in the dataset


The net total amount of "Profit/Losses" over the entire period


Calculate the changes in "Profit/Losses" over the entire period, then find the average of those changes


The greatest increase in profits (date and amount) over the entire period
The greatest decrease in losses (date and amount) over the entire 



==============================================================================================================================

The total number of votes cast


A complete list of candidates who received votes


The percentage of votes each candidate won


The total number of votes each candidate won


The winner of the election based on popular vote.